moved fd difference check to component
